requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.1
requests==2.31.0
gunicorn==21.2.0
